chore(refactor-code): refactor middleware code [Fix: #21993401]
