+++ server/app/middlewares/validateInput.js
@@ -3,28 +3,6 @@ import db from '../models/index';
 
 const Validate = {
 
-  /**
-   * Check for admin permission
-   * @param {Object} req - request object
-   * @param {any} res - response object
-   * @param {any} next - move to next controller handler
-   * @returns {void} no returns
-   */
-  hasAdminPermission(req, res, next) {
-    db.User
-      .findById(req.tokenDecode.userId)
-      .then((user) => {
-        if (user.roleId === 1) {
-          next();
-        } else {
-          return res.status(403)
-            .send({
-              message: 'You are not permitted to perform this action'
-            });
-        }
-      });
-  },
-
     /**
    * Validate user's input
    * @param {Object} req req object
@@ -185,31 +163,6 @@ const Validate = {
     next();
   },
 
-   /**
-   * Get a single user's profile
-   * @param {Object} req req object
-   * @param {Object} res response object
-   * @param {Object} next Move to next controller handler
-   * @returns {void|Object} response object or void
-   */
-  getUserName(req, res, next) {
-    db.User
-      .findOne({
-        where: { userName: req.query.q },
-      })
-      .then((user) => {
-        if (!user) {
-          return res.status(404)
-            .send({
-              message: 'This user does not exist'
-            });
-        }
-        req.getUser = user;
-        next();
-      })
-      .catch(err => res.status(500).send(err.errors));
-  },
-
   /**
    * Validate user to delete, make sure it not admin user
    * @param {Object} req req object
@@ -348,82 +301,6 @@ const Validate = {
     next();
   },
 
- /**
-   * Get a single user's profile
-   * @param {Object} req req object
-   * @param {Object} res response object
-   * @param {Object} next Move to next controller handler
-   * @returns {void|Object} response object or void
-   */
-  getSingleUser(req, res, next) {
-    db.User
-      .findOne({
-        where: { id: req.params.id },
-      })
-      .then((user) => {
-        if (!user) {
-          return res.status(404)
-            .send({
-              message: 'This user does not exist'
-            });
-        }
-        req.getUser = user;
-        next();
-      })
-      .catch(err => res.status(500).send(err.errors));
-  },
-
-  /**
-   * Checks if title is present in the request body
-   * @param {Object} req req object
-   * @param {Object} res response object
-   * @param {Object} next Move to next controller handler
-   * @returns {void|Object} response object or void
-   */
-  checkTitle(req, res, next) {
-    db.Role.findById(req.params.id)
-      .then(() => {
-        if (!req.body.title) {
-          return res.status(400)
-            .send({
-              message: 'Please input a value to update role with'
-            });
-        }
-        next();
-      });
-  },
-
- /**
-   * Get a single user's document
-   * @param {Object} req req object
-   * @param {Object} res response object
-   * @param {Object} next Move to next controller handler
-   * @returns {void|Object} response object or void
-   */
-  getSingleDocument(req, res, next) {
-    db.Document
-      .findById(req.params.id)
-      .then((document) => {
-        if (!document) {
-          return res.status(404)
-            .send({
-              message: 'This document cannot be found'
-            });
-        }
-        if (!Helper.isPublic(document) && !Helper.isOwnerDoc(document, req)
-           && !Helper.isAdmin(req.tokenDecode.roleId)
-           && !Helper.hasRoleAccess(document, req)) {
-          return res.status(401)
-            .send({
-              message: 'You are not permitted to view this document'
-            });
-        }
-        req.singleDocument = document;
-        next();
-      })
-      .catch(error => res.status(500).send(error.errors));
-  },
-
   /**
    * Validate documents input
    * @param {Object} req req object
@@ -473,87 +350,6 @@ const Validate = {
       OwnerRoleId: req.tokenDecode.roleId
     };
     next();
-  },
-
-  getDocumentByTitle(req, res, next) {
-    db.Document
-      .findAll({
-        where: { title: req.query.q },
-      })
-      .then((document) => {
-        if (!document) {
-          return res.status(404)
-            .send({
-              message: 'This document cannot be found'
-            });
-        }
-        if (!Helper.isPublic(document) && !Helper.isOwnerDoc(document, req)
-           && !Helper.isAdmin(req.tokenDecode.roleId)
-           && !Helper.hasRoleAccess(document, req)) {
-          return res.status(401)
-            .send({
-              message: 'You are not permitted to view this document'
-            });
-        }
-        req.singleDocument = document;
-        next();
-      })
-      .catch(error => res.status(500).send(error.errors));
-  },
-
- /**
-   * Check for document edit and delete permission
-   * @param {Object} req req object
-   * @param {Object} res response object
-   * @param {Object} next Move to next controller handler
-   * @returns {void|Object} response object or void
-   */
-  hasDocumentPermission(req, res, next) {
-    db.Document.findById(req.params.id)
-      .then((doc) => {
-        if (!doc) {
-          return res.status(404)
-            .send({
-              message: 'This document does not exist'
-            });
-        }
-        if (!Helper.isOwnerDoc(doc, req)
-          && !Helper.isAdmin(req.tokenDecode.roleId)) {
-          return res.status(401)
-            .send({
-              message: 'You are not permitted to modify this document'
-            });
-        }
-        req.docInstance = doc;
-        next();
-      });
-  },
-
-  /**
-   * Check for role edit and delete permission
-   * @param {Object} req req object
-   * @param {Object} res response object
-   * @param {Object} next Move to next controller handler
-   * @returns {void|Object} response object or void
-   */
-  modifyRolePermission(req, res, next) {
-    db.Role.findById(req.params.id)
-      .then((roles) => {
-        if (!roles) {
-          return res.status(404)
-            .send({
-              message: 'This role does not exist'
-            });
-        }
-        if (Helper.isAdmin(roles.id) || Helper.isRegular(roles.id)) {
-          return res.status(403)
-            .send({
-              message: 'You are not permitted to modify this role'
-            });
-        }
-        req.roleInstance = roles;
-        next();
-      });
   },
 };
